// frontend/src/app/core/services/api/model/accreditationStatus.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Enumeration of possible accreditation status values
 */
export const AccreditationStatus = {
    Active: 'ACTIVE',
    PendingValidation: 'PENDING_VALIDATION',
    PendingPhotoValidation: 'PENDING_PHOTO_VALIDATION',
    NewPhotoRequired: 'NEW_PHOTO_REQUIRED',
    Expired: 'EXPIRED',
    Suspended: 'SUSPENDED'
} as const;
export type AccreditationStatus = typeof AccreditationStatus[keyof typeof AccreditationStatus];